Decorator Changes (#18)
BREAKING CHANGE: 
Rename files interface.trace to interface.tracedocorator, the class name changed to InterfaceTraceDecorator ( it was Tracer not Trace)
changed return type of connect method to unique_ptr from raw pointer
changed passing and storing IInterface* to IInterface& - the implementation was never changed, the object may be recreated. If that is not the case it rather be shared/weak ptr (the class doesn't seems to own the IInterface class) (this enforces user to create the IInterfce as shared_ptr)
Removed AbstractDecorator class. Subscribing for interface changes is now in the InterfaceTraceDecorator - easier to keep track of the functions.