Align to changes in apigear-io/objectlink-core-cpp to v0.2.0 Rework the code for client and server.
Align to changes in apigear-io/objectlink-core-cpp to v0.2.0
Rework the code for both client and server. 

For both, extracted classes to separate files, add description of classes, clean up the interface if necessary. 
For code handling usage of WebSocket replaced POCO thread with std::async, remove inheritance of runnable. Add mutexes and locks where missing (change POCO mutex to std mutex). Add tests for basic setu and tear down scenarios from mermaid sequence diagrams (added to templates)

For client:
OlinkConnection(OlinkClient) class still takes care of reporting a sink to registry because it is one to know whether link remote needs to be called now (bugfix, for adding a sink after the olinkconnection was connected to host) or it will be called on connection established depending on connection state. for this purpose the IOlinkConnector was added. 
add possibility of sending messages right away (scenario of destroying olink connection - the unlink messages wouldn't be sent without it)
Align service adapter and remote interface classes, use IOlinkConnector in ctor and dtor, don't use the registry any more

For server:
Extracted OlinkWebSocketHandler handles request - either creates a websocket and passes to connection storage or fills the response (handling the not upgraded connection)
Extracted RequestHandlerFactory - only responsibility is providing a custom request handler: OlinkWebSocketHandler
Add ConnectionStorage as a class that keeps the connections, make it thread safe

Service adapter:
does not store node anymore- it was only a last added node, instead it asks registry for all the nodes it uses and sends them the signal changed or property changed
stores shared_ptr to interface, not a registry. It should co-own the service (interface impl)